+++ src/Modules/Admin/Controllers/Traits/TagTrait.php
@@ -5,21 +5,16 @@ declare(strict_types=1);
 namespace Hirtz\Location\Modules\Admin\Controllers\Traits;
 
 use Hirtz\Location\Models\Tag;
-use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 
 trait TagTrait
 {
-    protected function findTag(int $id, ?string $permissionName = null): Tag
+    protected function findTag(int $id): Tag
     {
         $tag = Tag::findOne($id);
 
         if (!$tag) {
             throw new NotFoundHttpException();
-        }
-
-        if ($permissionName && !$this->webuser->can($permissionName, ['tag' => $tag])) {
-            throw new ForbiddenHttpException();
         }
 
         return $tag;